NO AUTO Strips map, agg, convert, math, text, number from full
