+++ core/src/main/java/apoc/agg/CollAggregation.java
