+++ core/src/main/java/apoc/agg/Graph.java
