+++ core/src/main/java/apoc/agg/MaxAndMinItems.java
