+++ core/src/main/java/apoc/agg/Product.java
