+++ core/src/main/java/apoc/agg/Statistics.java
