+++ core/src/main/java/apoc/convert/Convert.java
@@ -1,29 +1,23 @@
 package apoc.convert;
 
 import apoc.coll.SetBackedList;
+import apoc.convert.utils.ConvertUtils;
 import apoc.meta.Meta.Types;
 import apoc.util.Util;
 import org.neo4j.graphdb.Entity;
 import org.neo4j.graphdb.Node;
 import org.neo4j.graphdb.Relationship;
-import org.neo4j.internal.helpers.collection.Iterators;
 import org.neo4j.logging.Log;
 import org.neo4j.procedure.Context;
 import org.neo4j.procedure.Description;
 import org.neo4j.procedure.Name;
 import org.neo4j.procedure.UserFunction;
 
-import java.lang.reflect.Array;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.Iterator;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-
 
 /**
  * @author mh
@@ -56,7 +50,7 @@ public class Convert {
     @UserFunction
     @Description("apoc.convert.toList(value) | tries it's best to convert the value to a list")
     public List<Object> toList(@Name("list") Object list) {
-        return convertToList(list);
+        return ConvertUtils.convertToList(list);
     }
 
     @UserFunction
@@ -78,33 +72,8 @@ public class Convert {
     }
 
     @SuppressWarnings("unchecked")
-    public static List convertToList(Object list) {
-        if (list == null) return null;
-        else if (list instanceof List) return (List) list;
-        else if (list instanceof Collection) return new ArrayList((Collection)list);
-        else if (list instanceof Iterable) return Iterators.addToCollection(((Iterable)list).iterator(),(List)new ArrayList<>(100));
-        else if (list instanceof Iterator) return Iterators.addToCollection((Iterator)list,(List)new ArrayList<>(100));
-        else if (list.getClass().isArray()) {
-            final Object[] objectArray;
-            if (list.getClass().getComponentType().isPrimitive()) {
-                int length = Array.getLength(list);
-                objectArray = new Object[length];
-                for (int i = 0; i < length; i++) {
-                    objectArray[i] = Array.get(list, i);
-                }
-            } else {
-                objectArray = (Object[]) list;
-            }
-            List result = new ArrayList<>(objectArray.length);
-            Collections.addAll(result, objectArray);
-            return result;
-        }
-        return Collections.singletonList(list);
-    }
-    
-    @SuppressWarnings("unchecked")
     private <T> List<T> convertToList(Object list, Class<T> type) {
-        List<Object> convertedList = convertToList(list);
+        List<Object> convertedList = ConvertUtils.convertToList(list);
         if (convertedList == null) {
         	return null;
         }
@@ -139,7 +108,7 @@ public class Convert {
     @UserFunction
     @Description("apoc.convert.toSet(value) | tries it's best to convert the value to a set")
     public List<Object> toSet(@Name("list") Object value) {
-        List list = convertToList(value);
+        List list = ConvertUtils.convertToList(value);
         return list == null ? null : new SetBackedList(new LinkedHashSet<>(list));
     }
