+++ core/src/main/java/apoc/convert/Json.java
@@ -1,6 +1,6 @@
 package apoc.convert;
 
-import apoc.meta.Meta;
+import apoc.convert.utils.ConvertUtils;import apoc.meta.Meta;
 import apoc.result.MapResult;
 import apoc.util.JsonUtil;
 import apoc.util.Util;
@@ -37,7 +37,7 @@ public class Json {
                         .map(i-> i instanceof Node ? nodeToMap((Node) i) : relToMap((Relationship) i))
                         .collect(Collectors.toList()));
             case LIST:
-                return Convert.convertToList(value).stream().map(Json::writeJsonResult).collect(Collectors.toList());
+                return ConvertUtils.convertToList(value).stream().map(Json::writeJsonResult).collect(Collectors.toList());
             case MAP:
                 return ((Map<String, Object>) value).entrySet()
                         .stream()
