+++ core/src/main/java/apoc/map/Maps.java
