+++ core/src/main/java/apoc/math/Maths.java
