+++ core/src/main/java/apoc/math/Regression.java
