+++ core/src/main/java/apoc/text/Phonetic.java
