+++ core/src/main/java/apoc/text/Strings.java
