+++ core/src/test/java/apoc/agg/CollAggregationTest.java
