+++ core/src/test/java/apoc/agg/MaxAndMinItemsAggregationTest.java
