+++ core/src/test/java/apoc/agg/ProductAggregationTest.java
