+++ core/src/test/java/apoc/agg/StatisticsTest.java
