+++ core/src/test/java/apoc/convert/ConvertTest.java
