+++ core/src/test/java/apoc/map/MapsTest.java
