+++ core/src/test/java/apoc/math/MathsTest.java
