+++ core/src/test/java/apoc/text/StringCleanTest.java
