+++ core/src/test/java/apoc/text/StringsTest.java
