+++ full/src/test/java/apoc/ApocSplitTest.java
@@ -91,7 +91,6 @@ public class ApocSplitTest {
         "apoc.atomic.insert",
         "apoc.atomic.remove",
         "apoc.atomic.update",
-        "apoc.math.regr",
         "apoc.search.nodeAllReduced",
         "apoc.search.nodeReduced",
         "apoc.search.multiSearchReduced",
@@ -110,7 +109,6 @@ public class ApocSplitTest {
         "apoc.schema.properties.distinct",
         "apoc.schema.properties.distinctCount",
         "apoc.log.stream",
-        "apoc.text.phoneticDelta",
         "apoc.spatial.sortByDistance",
         "apoc.spatial.geocodeOnce",
         "apoc.spatial.geocode",
@@ -137,8 +135,6 @@ public class ApocSplitTest {
         "apoc.warmup.run",
         "apoc.stats.degrees",
         "apoc.help",
-        "apoc.convert.setJsonProperty",
-        "apoc.convert.toTree",
         "apoc.neighbors.tohop",
         "apoc.neighbors.tohop.count",
         "apoc.neighbors.byhop",
@@ -204,112 +200,14 @@ public class ApocSplitTest {
         "apoc.hashing.fingerprint",
         "apoc.hashing.fingerprinting",
         "apoc.hashing.fingerprintGraph",
-        "apoc.math.maxLong",
-        "apoc.math.minLong",
-        "apoc.math.maxDouble",
-        "apoc.math.minDouble",
-        "apoc.math.maxInt",
-        "apoc.math.minInt",
-        "apoc.math.maxByte",
-        "apoc.math.minByte",
-        "apoc.math.sigmoid",
-        "apoc.math.sigmoidPrime",
-        "apoc.math.tanh",
-        "apoc.math.coth",
-        "apoc.math.cosh",
-        "apoc.math.sinh",
-        "apoc.math.sech",
-        "apoc.math.csch",
-        "apoc.number.format",
-        "apoc.number.parseInt",
-        "apoc.number.parseFloat",
-        "apoc.number.exact.add",
-        "apoc.number.exact.sub",
-        "apoc.number.exact.mul",
-        "apoc.number.exact.div",
-        "apoc.number.exact.toInteger",
-        "apoc.number.exact.toFloat",
-        "apoc.number.exact.toExact",
-        "apoc.number.romanToArabic",
-        "apoc.number.arabicToRoman",
         "apoc.schema.node.indexExists",
         "apoc.schema.relationship.indexExists",
         "apoc.schema.node.constraintExists",
         "apoc.schema.relationship.constraintExists",
         "apoc.xml.parse",
-        "apoc.map.groupBy",
-        "apoc.map.groupByMulti",
-        "apoc.map.fromNodes",
-        "apoc.map.fromPairs",
-        "apoc.map.fromLists",
-        "apoc.map.values",
-        "apoc.map.fromValues",
-        "apoc.map.merge",
-        "apoc.map.mergeList",
-        "apoc.map.get",
-        "apoc.map.mget",
-        "apoc.map.submap",
-        "apoc.map.setKey",
-        "apoc.map.setEntry",
-        "apoc.map.setPairs",
-        "apoc.map.setLists",
-        "apoc.map.setValues",
-        "apoc.map.removeKey",
-        "apoc.map.removeKeys",
-        "apoc.map.clean",
-        "apoc.map.updateTree",
-        "apoc.map.flatten",
-        "apoc.map.unflatten",
-        "apoc.map.sortedProperties",
         "apoc.version",
         "apoc.scoring.existence",
         "apoc.scoring.pareto",
-        "apoc.text.phonetic",
-        "apoc.text.doubleMetaphone",
-        "apoc.text.indexOf",
-        "apoc.text.indexesOf",
-        "apoc.text.replace",
-        "apoc.text.byteCount",
-        "apoc.text.bytes",
-        "apoc.text.regreplace",
-        "apoc.text.split",
-        "apoc.text.regexGroups",
-        "apoc.text.join",
-        "apoc.text.clean",
-        "apoc.text.compareCleaned",
-        "apoc.text.distance",
-        "apoc.text.levenshteinDistance",
-        "apoc.text.levenshteinSimilarity",
-        "apoc.text.hammingDistance",
-        "apoc.text.jaroWinklerDistance",
-        "apoc.text.sorensenDiceSimilarity",
-        "apoc.text.fuzzyMatch",
-        "apoc.text.urlencode",
-        "apoc.text.urldecode",
-        "apoc.text.lpad",
-        "apoc.text.rpad",
-        "apoc.text.format",
-        "apoc.text.slug",
-        "apoc.text.random",
-        "apoc.text.capitalize",
-        "apoc.text.capitalizeAll",
-        "apoc.text.decapitalize",
-        "apoc.text.decapitalizeAll",
-        "apoc.text.swapCase",
-        "apoc.text.camelCase",
-        "apoc.text.upperCamelCase",
-        "apoc.text.snakeCase",
-        "apoc.text.toUpperCase",
-        "apoc.text.base64Encode",
-        "apoc.text.base64Decode",
-        "apoc.text.base64UrlEncode",
-        "apoc.text.base64UrlDecode",
-        "apoc.text.charAt",
-        "apoc.text.code",
-        "apoc.text.hexValue",
-        "apoc.text.hexCharAt",
-        "apoc.text.toCypher",
-        "apoc.text.repeat",
         "apoc.bitwise.op",
         "apoc.data.url",
         "apoc.create.vNode",
@@ -318,39 +216,7 @@ public class ApocSplitTest {
         "apoc.create.uuid",
         "apoc.create.uuidBase64",
         "apoc.create.uuidBase64ToHex",
-        "apoc.create.uuidHexToBase64",
-        "apoc.json.path",
-        "apoc.convert.toJson",
-        "apoc.convert.getJsonProperty",
-        "apoc.convert.getJsonPropertyMap",
-        "apoc.convert.fromJsonMap",
-        "apoc.convert.fromJsonList",
-        "apoc.convert.toSortedJsonMap",
-        "apoc.convert.toMap",
-        "apoc.convert.toString",
-        "apoc.convert.toList",
-        "apoc.convert.toBoolean",
-        "apoc.convert.toNode",
-        "apoc.convert.toRelationship",
-        "apoc.convert.toSet",
-        "apoc.convert.toIntList",
-        "apoc.convert.toStringList",
-        "apoc.convert.toBooleanList",
-        "apoc.convert.toNodeList",
-        "apoc.convert.toRelationshipList",
-        "apoc.convert.toInteger",
-        "apoc.convert.toFloat",
-        "apoc.agg.percentiles",
-        "apoc.agg.product",
-        "apoc.agg.graph",
-        "apoc.agg.maxItems",
-        "apoc.agg.minItems",
-        "apoc.agg.statistics",
-        "apoc.agg.median",
-        "apoc.agg.nth",
-        "apoc.agg.first",
-        "apoc.agg.last",
-        "apoc.agg.slice");
+        "apoc.create.uuidHexToBase64");
 
     public static final List<String> FULL_PROCEDURES = List.of(
         "apoc.metrics.list",
